Modularize interpreter operations
