+++ js/interpreter.js
@@ -1,5 +1,7 @@
 // interpreter.js
 import { cities as sampleCities, people as samplePeople } from './samples.js';
+import { loadCsv, exportCsv } from './csv.js';
+import { keepColumns, withColumn, filterRows, joinDatasets } from './datasetOps.js';
 
 export class Interpreter {
     constructor(uiElements) {
@@ -95,32 +97,32 @@ export class Interpreter {
 
         switch (command) {
             case 'LOAD_CSV':
-                this.variables[this.activeVariableName] = await this.handleLoadCsv(args);
+                this.variables[this.activeVariableName] = await loadCsv(this, args);
                 break;
             case 'KEEP_COLUMNS':
             case 'SELECT':
                 if (!Array.isArray(currentDataset)) {
                     throw new Error(`No dataset loaded for VAR "${this.activeVariableName}" to apply ${command}.`);
                 }
-                this.variables[this.activeVariableName] = this.handleKeepColumns(args, currentDataset);
+                this.variables[this.activeVariableName] = keepColumns(this, args, currentDataset);
                 break;
             case 'JOIN':
                 if (!Array.isArray(currentDataset)) {
                     throw new Error(`No dataset loaded for VAR "${this.activeVariableName}" to apply JOIN.`);
                 }
-                this.variables[this.activeVariableName] = this.handleJoin(args, currentDataset);
+                this.variables[this.activeVariableName] = joinDatasets(this, args, currentDataset);
                 break;
             case 'FILTER':
                 if (!Array.isArray(currentDataset)) {
                     throw new Error(`No dataset loaded for VAR "${this.activeVariableName}" to apply FILTER.`);
                 }
-                this.variables[this.activeVariableName] = this.handleFilter(args, currentDataset);
+                this.variables[this.activeVariableName] = filterRows(this, args, currentDataset);
                 break;
             case 'WITH_COLUMN':
                 if (!Array.isArray(currentDataset)) {
                     throw new Error(`No dataset loaded for VAR "${this.activeVariableName}" to apply WITH_COLUMN.`);
                 }
-                this.variables[this.activeVariableName] = this.handleWithColumn(args, currentDataset);
+                this.variables[this.activeVariableName] = withColumn(this, args, currentDataset);
                 break;
             case 'PEEK':
                 // currentDataset is already what we want (array or null)
@@ -139,213 +141,10 @@ export class Interpreter {
                 if (!currentDataset) {
                     throw new Error(`No dataset available in VAR "${this.activeVariableName}" to export.`);
                 }
-                await this.handleExportCsv(args, currentDataset);
+                await exportCsv(this, args, currentDataset);
                 break;
             default: this.log(`Command ${command} for VAR "${this.activeVariableName}" is parsed but not yet fully implemented.`);
         }
     }
 
-    async handleLoadCsv(args) {
-        const fileName = args.file;
-        if (!fileName) throw new Error('LOAD_CSV requires FILE argument.');
-
-        if (typeof fetch !== 'undefined') {
-            try {
-                const resp = await fetch(`examples/${fileName}`);
-                if (resp.ok) {
-                    const text = await resp.text();
-                    return await this.parseCsvInput(text, fileName);
-                }
-            } catch (err) {
-                this.log(`Fetch for example ${fileName} failed: ${err.message}`);
-            }
-        }
-
-        if (!this.uiElements.csvFileInputEl) throw new Error('File input not available.');
-        const file = await this.requestCsvFile(fileName, this.activeVariableName);
-        return this.parseCsvInput(file, file.name);
-    }
-
-    parseCsvInput(input, name) {
-        return new Promise((resolve, reject) => {
-            this.log(`Using PapaParse for CSV parsing for VAR "${this.activeVariableName}".`);
-            if (typeof Papa === 'undefined') {
-                this.log('PapaParse library is not loaded.');
-                if (this.uiElements.fileInputContainerEl) this.uiElements.fileInputContainerEl.classList.add('hidden');
-                reject(new Error('PapaParse library is not available.'));
-                return;
-            }
-            Papa.parse(input, {
-                header: true,
-                skipEmptyLines: true,
-                dynamicTyping: true,
-                complete: (results) => {
-                    this.log(`Loaded ${results.data.length} rows for VAR "${this.activeVariableName}" from ${name}. Headers: ${results.meta.fields ? results.meta.fields.join(', ') : 'N/A'}`);
-                    if (this.uiElements.fileInputContainerEl) this.uiElements.fileInputContainerEl.classList.add('hidden');
-                    const rows = results.data;
-                    this.log(`Parsed CSV for VAR "${this.activeVariableName}". Rows: ${rows.length}`);
-                    resolve(rows);
-                },
-                error: (err) => {
-                    this.log(`PapaParse error for VAR "${this.activeVariableName}": ${err.message}`);
-                    if (this.uiElements.fileInputContainerEl) this.uiElements.fileInputContainerEl.classList.add('hidden');
-                    reject(err);
-                }
-            });
-        });
-    }
-
-    handleKeepColumns(args, currentDataset) {
-        // currentDataset is an array of objects
-        const { columns } = args;
-        if (!Array.isArray(columns)) {
-            throw new Error(`Invalid columns argument for KEEP_COLUMNS in VAR "${this.activeVariableName}".`);
-        }
-        if (!Array.isArray(currentDataset) || currentDataset.length === 0) {
-            return [];
-        }
-        const allCols = Object.keys(currentDataset[0]);
-        const columnsToKeep = columns.map(c => allCols.find(ac => ac.toLowerCase() === c.toLowerCase())).filter(Boolean);
-
-        if (columnsToKeep.length === 0) {
-            throw new Error(`None of the specified columns for KEEP_COLUMNS were found in VAR "${this.activeVariableName}".`);
-        }
-
-        const newDataset = currentDataset.map(row => {
-            const obj = {};
-            columnsToKeep.forEach(col => { obj[col] = row[col]; });
-            return obj;
-        });
-        this.log(`Kept columns: ${columnsToKeep.join(', ')} for VAR "${this.activeVariableName}".`);
-        return newDataset;
-    }
-
-    handleWithColumn(args, currentDataset) {
-        const { columnName, expression } = args;
-        if (!Array.isArray(expression) || expression.length === 0) {
-            throw new Error('WITH_COLUMN requires an expression.');
-        }
-
-        const evalExpr = (row) => {
-            const exprStr = expression.map(t => {
-                if (t.type === 'IDENTIFIER') return `row["${t.value}"]`;
-                if (t.type === 'NUMBER_LITERAL') return t.value;
-                if (t.type === 'OPERATOR' || (t.type === 'PUNCTUATION' && ['(', ')'].includes(t.value))) return t.value;
-                throw new Error(`Unsupported token ${t.value} in expression`);
-            }).join(' ');
-            try {
-                return Function('row', `return ${exprStr}`)(row);
-            } catch (e) {
-                throw new Error(`Error evaluating expression '${exprStr}': ${e.message}`);
-            }
-        };
-
-        const result = currentDataset.map(row => ({ ...row, [columnName]: evalExpr(row) }));
-        this.log(`WITH_COLUMN '${columnName}' computed for VAR "${this.activeVariableName}".`);
-        return result;
-    }
-
-
-    handleFilter(condition, currentDataset) {
-        const evalCondition = (node, row) => {
-            if (!node) return false;
-            if (node.type === 'AND') {
-                return evalCondition(node.left, row) && evalCondition(node.right, row);
-            }
-            if (node.type === 'OR') {
-                return evalCondition(node.left, row) || evalCondition(node.right, row);
-            }
-            const { column, operator = '=', value } = node;
-            const getVal = () => {
-                if (value && typeof value === 'object' && value.type === 'COLUMN_REFERENCE') {
-                    return row[value.name];
-                }
-                return value;
-            };
-            const a = row[column];
-            const b = getVal();
-            switch (operator) {
-                case '=':
-                case 'IS':
-                    return a === b;
-                case 'IS NOT':
-                    return a !== b;
-                case '!=':
-                    return a != b;
-                case '>':
-                    return a > b;
-                case '<':
-                    return a < b;
-                case '>=':
-                    return a >= b;
-                case '<=':
-                    return a <= b;
-                case 'CONTAINS':
-                    return String(a).includes(String(b));
-                case 'STARTSWITH':
-                    return String(a).startsWith(String(b));
-                case 'ENDSWITH':
-                    return String(a).endsWith(String(b));
-                default:
-                    throw new Error(`Unsupported operator ${operator}`);
-            }
-        };
-
-        const filtered = currentDataset.filter(row => evalCondition(condition, row));
-        this.log(`FILTER kept ${filtered.length} of ${currentDataset.length} rows for VAR "${this.activeVariableName}".`);
-        return filtered;
-    }
-
-    handleJoin(args, currentDataset) {
-        const { variable, leftKey, rightKey, type = 'INNER' } = args;
-        const other = this.variables[variable];
-        if (!Array.isArray(other)) {
-            throw new Error(`JOIN target VAR "${variable}" is not loaded or not an array.`);
-        }
-        if (!Array.isArray(currentDataset)) {
-            throw new Error(`Current dataset for VAR "${this.activeVariableName}" is not an array.`);
-        }
-
-        const map = new Map();
-        for (const row of other) {
-            if (row.hasOwnProperty(rightKey)) {
-                const key = row[rightKey];
-                if (!map.has(key)) map.set(key, []);
-                map.get(key).push(row);
-            }
-        }
-
-        const joined = [];
-        for (const lRow of currentDataset) {
-            const key = lRow[leftKey];
-            const matches = map.get(key);
-            if (matches) {
-                for (const rRow of matches) {
-                    joined.push({ ...lRow, ...rRow });
-                }
-            } else if (type === 'LEFT') {
-                joined.push({ ...lRow });
-            }
-        }
-        this.log(`JOIN ${type} completed using '${leftKey}' = '${rightKey}' with VAR "${variable}". Rows: ${joined.length}`);
-        return joined;
-    }
-
-    async handleExportCsv(args, dataset) {
-        const fileName = args.file || 'export.csv';
-
-        if (Array.isArray(dataset) && dataset.length > 0 && typeof dataset[0] === 'object') {
-            const csvString = Papa.unparse(dataset);
-            const blob = new Blob([csvString], { type: 'text/csv;charset=utf-8;' });
-            const link = document.createElement('a');
-            link.href = URL.createObjectURL(blob);
-            link.download = fileName;
-            document.body.appendChild(link);
-            link.click();
-            document.body.removeChild(link);
-            this.log(`Exported data from VAR "${this.activeVariableName}" to ${fileName}.`);
-        } else {
-            throw new Error(`EXPORT_CSV does not support dataset type: ${typeof dataset}`);
-        }
-    }
 }
